MakieExtras

Documentation for MakieExtras.

MakieExtras.plotseries!

MakieExtras.plotseries!

MakieExtras.set_trackpad_pan!

MakieExtras.plotseries! — Method

Plot timeseries from rows of a matrix x, with time vector t. Each series is plotted in a different color. A legend and data inspector are added to the axis.

MakieExtras.plotseries! — Method

plotseries! with default arguments for a new figure and axis.

MakieExtras.set_trackpad_pan! — Method

Set up an axis for trackpad panning. This will disable the default scroll-to-zoom and instead allow you to pan the axis by dragging with two fingers on a trackpad. Zooming with two fingers will work with the shift key pressed.